feat: add appendPriorityArray, add decimal action
Create appendPriorityArray so that priorityArray will be updated based
off the lookup array. Add appendPriorityArray to operator switch cases.

Move decimal case to the other standard number input cases, to reduce
redundant code. Add condition for decimal to not allow decimal to be
inputted if there already is a decimal in the display.